Use Long for Timestamp.nanos

+++ src/commonMain/kotlin/klogger/Timestamp.kt
@@ -2,19 +2,16 @@ package klogger
 
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.builtins.serializer
 import kotlinx.serialization.descriptors.PrimitiveKind
 import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
 import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.descriptors.buildClassSerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
-import kotlinx.serialization.encoding.encodeStructure
 
 expect fun now(): Timestamp
 
 @Serializable(with = TimestampSerialiser::class)
-data class Timestamp(val epochSeconds: Long, val nanos: Int) {
+data class Timestamp(val epochSeconds: Long, val nanos: Long) {
     override fun toString(): String {
         val ns = "000000000$nanos"
         return "$epochSeconds.${ns.substring(ns.length - 9)}"
@@ -31,7 +28,7 @@ object TimestampSerialiser : KSerializer<Timestamp> {
 
     override fun deserialize(decoder: Decoder): Timestamp {
         val (epochSeconds, nanos) = decoder.decodeString().split(".")
-        return Timestamp(epochSeconds.toLong(), nanos.toInt())
+        return Timestamp(epochSeconds.toLong(), nanos.toLong())
     }
 }
