+++ src/jvmMain/kotlin/klogger/now.kt
@@ -4,5 +4,5 @@ import java.time.Instant
 
 actual fun now(): Timestamp {
     val instant = Instant.now()
-    return Timestamp(instant.epochSecond, instant.nano)
+    return Timestamp(instant.epochSecond, instant.nano.toLong())
 }
